safari-534.53-branch/LayoutTests/inspector/console/console-assert-expected.txt

CONSOLE MESSAGE: line 9: 1
CONSOLE MESSAGE: line 10: a
Tests that console.assert() will dump a message and stack trace with source URLs and line numbers.

console-assert.html:9Assertion failed: 1bconsole-assert.html:9aconsole-assert.html:15
console-assert.html:10Assertion failed: a bbconsole-assert.html:10aconsole-assert.html:15
